Debris is falling
The debris is now falling down the screen. All of the pieces of the
debris are falling down at the same time and have no hit detection. My
goal for next class is to get the hit detection working and maybe get
the debris to fall at different times and for them to go back to the
top of the screen after falling off of it.